feat(Table): UXD-1313 moving Headers as its own component
