+++ packages/Table/src/Table.styles.js
@@ -1,6 +1,6 @@
 import styled, { css } from "styled-components";
 import tokens from "@paprika/tokens";
-import { fontSize } from "@paprika/stylers/lib/helpers";
+
 import * as types from "./types";
 
 export const Table = styled.table(() => {
@@ -13,7 +13,7 @@ export const Table = styled.table(() => {
     }
 
     & .is-highlighted-focus {
-      outline: 2px solid blue;
+      outline: 2px solid ${tokens.highlight.active.noBorder.borderColor};
     }
     & .is-highlighted-idle {
       outline: 2px solid transparent;
@@ -39,13 +39,7 @@ export const TBody = styled.tbody(({ hasZebraStripes }) => {
   `;
 });
 
-export const Thead = styled.thead`
-  background: ${tokens.table.header.backgroundColor};
-  position: relative;
-  text-align: left;
-`;
-
-const borderTypesStyles = {
+export const borderTypesStyles = {
   [types.NONE]: "",
   [types.VERTICAL]: css`
     border-color: ${tokens.border.color};
@@ -71,23 +65,5 @@ export const TD = styled.td(({ borderType, cellPropsResetCSS, width = null, stic
     ${width ? `width: ${width}${px}; max-width: ${width}${px};` : ""};
     text-align: left;
     z-index: 1;
-  `;
-});
-
-export const TH = styled.th(({ borderType, sticky }) => {
-  return css`
-    ${fontSize()}
-    ${borderType in borderTypesStyles ? borderTypesStyles[borderType] : ""}
-    background: ${tokens.table.header.backgroundColor};
-    font-weight: bold;
-    padding: ${tokens.space};
-    position: sticky;
-    text-align: left;
-    /** [Sticky] This will work only when the header is made of one TR more work will be need to do
-    in order to support complex headers with multiple TH
-    */
-    top: -1px;
-    z-index: 2;
-    ${typeof sticky !== "undefined" ? `z-index: 3; left: ${!Number.isNaN(sticky) ? sticky : 0}px;` : ""};
   `;
 });
