--- packages/Table/src/components/Headers.js
@@ -1,6 +1,0 @@
-export default function Headers(props) {
-  const { children } = props;
-  return children;
-}
-
-Headers.displayName = "Table.Headers";
