+++ packages/Table/src/components/Headers/index.js
@@ -1,0 +1,1 @@
+export { default, Headers } from "./Headers";
